Update cachematrix.R
I edited the code to create the makeCacheMatrix that that creates a  "matrix" object that can cache its inverse for the input that is an invertible square matrix. 
After, I made the cacheSolve is a function that computes the inverse of the special "matrix" created by the makeCacheMatrix function created previously. If the inverse has already been calculated and the matrix has not changed, then the cachesolve function should retrieve the inverse from the cache.